chore: reuse getMetaTagsContentByName
